* 5.4:
  [HttpFoundation] IPv4-mapped IPv6 addresses incorrectly rejected
  [RateLimiter] Add typecast to Reservation::wait
  [VarDumper] Ignore \Error in __debugInfo()

+++ src/Symfony/Component/RateLimiter/Reservation.php
@@ -45,6 +45,6 @@ final class Reservation
 
     public function wait(): void
     {
-        usleep($this->getWaitDuration() * 1e6);
+        usleep((int) ($this->getWaitDuration() * 1e6));
     }
 }
